+++ packages/tentales/src/middleware/middlewares/__tests__/body-parser-middleware.test.ts
@@ -1,7 +1,7 @@
 import Koa from "koa"
 import supertest from "supertest"
-import { bodyParserMiddleware } from "./body-parser-middleware"
-import { createTestMw, createTestMwConfig } from "./__test-helpers__"
+import { bodyParserMiddleware } from "../body-parser-middleware"
+import { createTestMw, createTestMwConfig } from "../__test-helpers__"
 
 describe("Incoming request", () => {
   test("It does parse the body", async done => {
